Fix tests after directory restructuring
